update(api): move /ping to HealthController
