+++ packages/api/core/src/__tests__/health.rs
@@ -1,0 +1,20 @@
+use crate::__tests__::prelude::*;
+
+#[sqlx::test]
+async fn ping(pool: PgPool) -> Result<()> {
+    // Arrange
+    let (server, _)= init(pool).await;
+
+    // Act
+    let response = server.get("/ping").await;
+
+    // Assert
+    response.assert_status_ok();
+
+    response.assert_json(&json!({
+        "success": true,
+        "message": "pong",
+    }));
+
+    Ok(())
+}
