+++ packages/api/core/src/http/controllers/mod.rs
@@ -1,16 +1,16 @@
 mod auth;
-mod dev;
 mod errors;
 mod exercise;
 mod exercise_equipment;
 mod link;
 mod muscle;
 mod muscle_group;
+mod health;
 
 pub use auth::AuthController;
-pub use dev::DevController;
 pub use exercise::ExerciseController;
 pub use exercise_equipment::ExerciseEquipmentController;
+pub use health::HealthController;
 pub use link::LinkController;
 pub use muscle::MuscleController;
 pub use muscle_group::MuscleGroupController;
